사용자 회원가입 기능 추가(위경도, 프로필 이미지 저장)

+++ src/main/java/com/javaex/vo/AddressVo.java
@@ -8,16 +8,15 @@ public class AddressVo {
 	private String address_name;
 	private String main_address;
 	private String sub_address;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	private int default_address;
 	
-	
-	//생성자
+	// 생성자
 	public AddressVo() {}
-
-	public AddressVo(int address_no, int no, String address_name, String main_address, String sub_address, int lat,
-			int lng, int default_address) {
+	
+	public AddressVo(int address_no, int no, String address_name, String main_address, String sub_address, double lat,
+			double lng, int default_address) {
 		this.address_no = address_no;
 		this.no = no;
 		this.address_name = address_name;
@@ -33,86 +32,72 @@ public class AddressVo {
 		return address_no;
 	}
 
-
 	public void setAddress_no(int address_no) {
 		this.address_no = address_no;
 	}
-
 
 	public int getNo() {
 		return no;
 	}
 
-
 	public void setNo(int no) {
 		this.no = no;
 	}
-
 
 	public String getAddress_name() {
 		return address_name;
 	}
 
-
 	public void setAddress_name(String address_name) {
 		this.address_name = address_name;
 	}
-
 
 	public String getMain_address() {
 		return main_address;
 	}
 
-
 	public void setMain_address(String main_address) {
 		this.main_address = main_address;
 	}
-
 
 	public String getSub_address() {
 		return sub_address;
 	}
 
-
 	public void setSub_address(String sub_address) {
 		this.sub_address = sub_address;
 	}
 
-
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
-
 
 	public int getDefault_address() {
 		return default_address;
 	}
 
-	
 	public void setDefault_address(int default_address) {
 		this.default_address = default_address;
 	}
 
-	// 일반 메소드
+	// 메소드 일반
 	@Override
 	public String toString() {
 		return "AddressVo [address_no=" + address_no + ", no=" + no + ", address_name=" + address_name
 				+ ", main_address=" + main_address + ", sub_address=" + sub_address + ", lat=" + lat + ", lng=" + lng
 				+ ", default_address=" + default_address + "]";
-	}
+	}	
+	
 }
